Refactoring - moved classes to their own files and renaming
