+++ src/Dashboard/App_Start/Bootstrap.cs
@@ -28,7 +28,7 @@ namespace Dashboard
                 init.For<IRepository>().Use(c => _repository);
                 init.For<IApplicationSettings>().Use(() => _applicationSettings);
                 init.For<IUniqueKeyGenerator>().Use<UniqueKeyGenerator>();
-                init.For<IDashboardProvider>().Use<RAGWidgetViewProvider>();
+                init.For<IRAGWidgetViewProvider>().Use<RAGWidgetViewProvider>();
                 init.For<IProjectionWriter<RAGWidgetView>>().Use<MongoDbProjectionWriter<RAGWidgetView>>().Ctor<string>("connectionString").Is(_applicationSettings.MongoDbConnectionString).Ctor<string>("databaseName").Is(_applicationSettings.MongoDbName);
                 init.For<ILog>().Singleton().Use(c => LogManager.GetLogger("Dashboard"));
             });
