+++ src/Dashboard/Dashboard.csproj
@@ -128,7 +128,11 @@
     <Compile Include="Contracts\Commands\DashboardCommands.cs" />
     <Compile Include="Contracts\Events\DashboardEvents.cs" />
     <Compile Include="Controllers\AopController.cs" />
+    <Compile Include="Models\RAGWidgetModels\CreateRAGWidgetRequest.cs" />
     <Compile Include="Controllers\DashboardController.cs" />
+    <Compile Include="Models\RAGWidgetModels\TextValue.cs" />
+    <Compile Include="Models\RAGWidgetModels\RAGWidgetInfo.cs" />
+    <Compile Include="Extentions\ExtendsRAGWidgetInfo.cs" />
     <Compile Include="Domain\Base\AggregateBase.cs" />
     <Compile Include="Domain\Base\IAggregate.cs" />
     <Compile Include="Domain\Base\IAggregateFactory.cs" />
