+++ src/Dashboard/Models/RAGWidgetModels/CreateRAGWidgetRequest.cs
@@ -1,0 +1,7 @@
+﻿namespace Dashboard.Models.RAGWidgetModels
+{
+    public class CreateRAGWidgetRequest
+    {
+        public string Name { get; set; }
+    }
+}
