+++ src/Dashboard/Models/RAGWidgetModels/RAGWidgetInfo.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+using Dashboard.Controllers;
+
+namespace Dashboard.Models.RAGWidgetModels
+{
+    public class RAGWidgetInfo
+    {
+        public List<TextValue> item { get; set; }
+    }
+}
