+++ src/Dashboard/Models/RAGWidgetModels/TextValue.cs
@@ -1,0 +1,8 @@
+namespace Dashboard.Models.RAGWidgetModels
+{
+    public class TextValue
+    {
+        public string text { get; set; }
+        public int value { get; set; }
+    }
+}
