+++ src/Dashboard/ReadModel/Providers/RAGWidgetViewProvider.cs
@@ -1,20 +1,18 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using System.Web;
 using Dashboard.Infrastructure;
 using Dashboard.ReadModel.Views;
 using MongoDB.Driver;
 
 namespace Dashboard.ReadModel.Providers
 {
-    public interface IDashboardProvider : IProvider<RAGWidgetView>
+    public interface IRAGWidgetViewProvider : IProvider<RAGWidgetView>
     {
         Task<RAGWidgetView> Get(Guid id);
     }
 
-    public class RAGWidgetViewProvider : MongoDbProvider<RAGWidgetView>, IDashboardProvider
+    public class RAGWidgetViewProvider : MongoDbProvider<RAGWidgetView>, IRAGWidgetViewProvider
     {
         public RAGWidgetViewProvider(IApplicationSettings applicationSettings) : base(applicationSettings)
         {}
